Revised GCode Extraction through mapping function
If an invalid variable is passed in from the string,the function will 
loop according to the length of the string and will check whether it has 
reached the end of the string after exitting the for loop.
If it has yet to reach the end,an error will be thrown indicating the 
variable does not exist in this command.